+++ lib/open_meteo/types.rb
@@ -1,0 +1,10 @@
+require "dry-types"
+require "dry-struct"
+
+module OpenMeteo
+  ##
+  # The types from the dry-types gem.
+  module Types
+    include Dry.Types()
+  end
+end